fix(embed): fetch project data outside the details pane
- Move the project fetch and its error redirect into `useInitializeSingleProject`, which runs above the layout gate
- Restores interventions, time travel config and error handling for `embed=true` with `project_details=false`
- Key the fetch on the `p` route param rather than the asynchronously derived page, so it does not wait a render
