+++ src/hooks/useInitializeSingleProject.ts
@@ -3,12 +3,16 @@ import type { ProjectSiteFeature } from '../features/common/types/map';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import {
+  useCurrencyStore,
   useInterventionStore,
   useSingleProjectStore,
+  useTenantStore,
   useViewStore,
 } from '../stores';
 import { useLocale } from 'next-intl';
 import { FIRST_SITE_INDEX, hasNoSites, isString } from '../utils/projectV2';
+import { useApi } from './useApi';
+import useLocalizedPath from './useLocalizedPath';
 
 const getSiteIndexById = (
   sites: ProjectSiteFeature[],
@@ -22,24 +26,74 @@ const getSiteIndexById = (
 export const useInitializeSingleProject = () => {
   const locale = useLocale();
   const router = useRouter();
-  const { site: requestedSite, ploc: requestedIntervention } = router.query;
+  const { getApi } = useApi();
+  const { localizedPath } = useLocalizedPath();
+  const {
+    p: projectSlug,
+    site: requestedSite,
+    ploc: requestedIntervention,
+  } = router.query;
   const hasOnlyRequestedIntervention = Boolean(
     !requestedSite && requestedIntervention
   );
   // store: state
   const currentPage = useViewStore((state) => state.page);
+  const currencyCode = useCurrencyStore((state) => state.currencyCode);
+  const tenantId = useTenantStore((state) => state.tenantConfig.id);
   const singleProject = useSingleProjectStore((state) => state.singleProject);
   const selectedSite = useSingleProjectStore((state) => state.selectedSite);
+  const fetchError = useSingleProjectStore((state) => state.fetchError);
   const selectedIntervention = useInterventionStore(
     (state) => state.selectedIntervention
   );
   // store: action
+  const fetchProject = useSingleProjectStore((state) => state.fetchProject);
   const selectSiteAndSyncUrl = useSingleProjectStore(
     (state) => state.selectSiteAndSyncUrl
   );
   const clearProjectStates = useSingleProjectStore(
     (state) => state.clearProjectStates
   );
+
+  /**
+   * Fetch the project whenever the URL names one.
+   *
+   * This lives here rather than in `ProjectDetails` because that component is
+   * not rendered when the details pane is hidden (`embed=true` with
+   * `project_details=false`), which used to leave the map, interventions and
+   * time travel config with no data at all.
+   *
+   * `router.query.p` is used as the signal rather than `viewStore.page`, since
+   * the page is derived from that same param but only after an effect, and this
+   * fetch should not wait on it.
+   */
+  useEffect(() => {
+    if (!router.isReady) return;
+    if (!isString(projectSlug)) return;
+    if (!currencyCode) return;
+
+    fetchProject(
+      getApi,
+      {
+        queryParams: {
+          _scope: 'extended',
+          currency: currencyCode,
+          //passing locale/tenant as a query param to break cache when locale changes,
+          //as the browser uses the cached response even though the x-locale header is different
+          locale,
+          tenant: tenantId,
+        },
+      },
+      projectSlug
+    );
+  }, [router.isReady, projectSlug, locale, currencyCode, tenantId]);
+
+  // Redirect home if the project fails to load. The store handles the error message, and this prevents an endless loading state.
+  useEffect(() => {
+    if (fetchError) {
+      router.push(localizedPath('/'));
+    }
+  }, [fetchError]);
 
   const projectSites = singleProject?.sites ?? [];
   // A project has no sites if all site geometries are null.
